Fix errors in IDL, and stagger the codes across programs (#636)
