+++ chains/solana/contracts/programs/ccip-offramp/src/instructions/v1/ocr3base.rs
@@ -2,8 +2,9 @@ use anchor_lang::prelude::*;
 use anchor_lang::solana_program::sysvar;
 use anchor_lang::solana_program::{keccak, secp256k1_recover::*};
 
-use crate::ocr3base::{ConfigSet, Ocr3Error, Transmitted, MAX_ORACLES};
+use crate::ocr3base::{ConfigSet, Transmitted, MAX_ORACLES};
 use crate::state::{Ocr3Config, Ocr3ConfigInfo};
+use crate::CcipOfframpError;
 
 pub const MAX_SIGNERS: usize = MAX_ORACLES;
 pub const MAX_TRANSMITTERS: usize = MAX_ORACLES;
@@ -55,9 +56,12 @@ pub fn ocr3_set(
 ) -> Result<()> {
     require!(
         plugin_type == ocr3_config.plugin_type,
-        Ocr3Error::InvalidPluginType
+        CcipOfframpError::Ocr3InvalidPluginType
     );
-    require!(cfg.f != 0, Ocr3Error::InvalidConfigFMustBePositive);
+    require!(
+        cfg.f != 0,
+        CcipOfframpError::Ocr3InvalidConfigFMustBePositive
+    );
 
     // If F is 0, then the config is not yet set
     if ocr3_config.config_info.f == 0 {
@@ -67,13 +71,13 @@ pub fn ocr3_set(
         require!(
             ocr3_config.config_info.is_signature_verification_enabled
                 == cfg.is_signature_verification_enabled,
-            Ocr3Error::StaticConfigCannotBeChanged
+            CcipOfframpError::Ocr3StaticConfigCannotBeChanged
         )
     };
 
     require!(
         transmitters.len() <= MAX_TRANSMITTERS,
-        Ocr3Error::InvalidConfigTooManyTransmitters
+        CcipOfframpError::Ocr3InvalidConfigTooManyTransmitters
     );
 
     clear_transmitters(ocr3_config);
@@ -82,11 +86,11 @@ pub fn ocr3_set(
         clear_signers(ocr3_config);
         require!(
             signers.len() <= MAX_SIGNERS,
-            Ocr3Error::InvalidConfigTooManySigners
+            CcipOfframpError::Ocr3InvalidConfigTooManySigners
         );
         require!(
             signers.len() > 3 * cfg.f as usize,
-            Ocr3Error::InvalidConfigFIsTooHigh
+            CcipOfframpError::Ocr3InvalidConfigFIsTooHigh
         );
 
         ocr3_config.config_info.n = signers.len() as u8;
@@ -135,7 +139,7 @@ pub(super) fn ocr3_transmit<R: Ocr3Report>(
 ) -> Result<()> {
     require!(
         plugin_type == ocr3_config.plugin_type,
-        Ocr3Error::InvalidPluginType
+        CcipOfframpError::Ocr3InvalidPluginType
     );
 
     // validate raw message length
@@ -157,12 +161,12 @@ pub(super) fn ocr3_transmit<R: Ocr3Report>(
     require_eq!(
         tx.data.len() as u128,
         expected_data_len,
-        Ocr3Error::WrongMessageLength
+        CcipOfframpError::Ocr3WrongMessageLength
     );
 
     require!(
         ocr3_config.config_info.config_digest == report_context.byte_words[0],
-        Ocr3Error::ConfigDigestMismatch
+        CcipOfframpError::Ocr3ConfigDigestMismatch
     );
 
     // TODO: chain fork check - is this even possible?
@@ -171,18 +175,18 @@ pub(super) fn ocr3_transmit<R: Ocr3Report>(
     ocr3_config
         .transmitters
         .binary_search_by(|probe| transmitter.to_bytes().cmp(probe))
-        .map_err(|_| Ocr3Error::UnauthorizedTransmitter)?;
+        .map_err(|_| CcipOfframpError::Ocr3UnauthorizedTransmitter)?;
 
     if ocr3_config.config_info.is_signature_verification_enabled != 0 {
         require_eq!(
             signatures.rs.len(),
             (ocr3_config.config_info.f + 1) as usize,
-            Ocr3Error::WrongNumberOfSignatures
+            CcipOfframpError::Ocr3WrongNumberOfSignatures
         );
         require_eq!(
             signatures.rs.len(),
             signatures.ss.len(),
-            Ocr3Error::SignaturesOutOfRegistration
+            CcipOfframpError::Ocr3SignaturesOutOfRegistration
         );
 
         verify_signatures(ocr3_config, report.hash(&report_context), signatures)?;
@@ -211,22 +215,22 @@ fn verify_signatures(
             signatures.raw_vs[i],
             [signatures.rs[i], signatures.ss[i]].concat().as_ref(),
         )
-        .map_err(|_| Ocr3Error::InvalidSignature)?;
+        .map_err(|_| CcipOfframpError::Ocr3InvalidSignature)?;
         // convert to a raw 20 byte Ethereum address
         let address: [u8; 20] = keccak::hash(&signer.0).to_bytes()[12..32]
             .try_into()
-            .map_err(|_| Ocr3Error::UnauthorizedSigner)?;
+            .map_err(|_| CcipOfframpError::Ocr3UnauthorizedSigner)?;
         let index = ocr3_config
             .signers
             .binary_search_by(|probe| address.cmp(probe))
-            .map_err(|_| Ocr3Error::UnauthorizedSigner)?;
+            .map_err(|_| CcipOfframpError::Ocr3UnauthorizedSigner)?;
 
         uniques |= 1 << index;
     }
 
     require!(
         uniques.count_ones() as usize == signatures.rs.len(),
-        Ocr3Error::NonUniqueSignatures
+        CcipOfframpError::Ocr3NonUniqueSignatures
     );
 
     Ok(())
@@ -237,10 +241,16 @@ fn assign_oracles<const A: usize>(oracles: &mut [[u8; A]], location: &mut [[u8; 
     // sort + verify not duplicated
     oracles.sort_unstable_by(|a, b| b.cmp(a)); // reverse sort to ensure appended 0-values in correct
     let duplicate = oracles.windows(2).any(|pair| pair[0] == pair[1]);
-    require!(!duplicate, Ocr3Error::InvalidConfigRepeatedOracle);
+    require!(
+        !duplicate,
+        CcipOfframpError::Ocr3InvalidConfigRepeatedOracle
+    );
 
     for (i, o) in oracles.iter().enumerate() {
-        require!(*o != [0; A], Ocr3Error::OracleCannotBeZeroAddress);
+        require!(
+            *o != [0; A],
+            CcipOfframpError::Ocr3OracleCannotBeZeroAddress
+        );
         location[i] = *o;
     }
     Ok(())
