+++ chains/solana/contracts/programs/fee-quoter/src/lib.rs
@@ -315,18 +315,12 @@ pub mod fee_quoter {
     }
 }
 
-// TODO this is a hack because Anchor + Anchor-Go fail to include all errors in the IDL and the gobindings.
-// By having this first (though unused) error enum here, it does pick up the actual (second) error enum
-#[error_code]
-pub enum AnchorErrorHack {
-    Something,
-    Else,
-}
-
 #[error_code]
 pub enum FeeQuoterError {
     #[msg("The signer is unauthorized")]
-    Unauthorized,
+    // offset error code so that they don't clash with other programs
+    // (Anchor's base custom error code 6000 + offset 2000 = start at 8000)
+    Unauthorized = 2000,
     #[msg("Invalid inputs")]
     InvalidInputs,
     #[msg("Gas limit is zero")]
