+++ chains/solana/gobindings/ccip_offramp/types.go
@@ -846,126 +846,3 @@ func (value CodeVersion) String() string {
 		return ""
 	}
 }
-
-type CcipOfframpError ag_binary.BorshEnum
-
-const (
-	InvalidSequenceInterval_CcipOfframpError CcipOfframpError = iota
-	RootNotCommitted_CcipOfframpError
-	ExistingMerkleRoot_CcipOfframpError
-	Unauthorized_CcipOfframpError
-	InvalidNonce_CcipOfframpError
-	InvalidInputsMissingWritable_CcipOfframpError
-	OnrampNotConfigured_CcipOfframpError
-	FailedToDeserializeReport_CcipOfframpError
-	InvalidPluginType_CcipOfframpError
-	InvalidVersion_CcipOfframpError
-	MissingExpectedPriceUpdates_CcipOfframpError
-	MissingExpectedMerkleRoot_CcipOfframpError
-	UnexpectedMerkleRoot_CcipOfframpError
-	RedundantOwnerProposal_CcipOfframpError
-	UnsupportedSourceChainSelector_CcipOfframpError
-	UnsupportedDestinationChainSelector_CcipOfframpError
-	InvalidProof_CcipOfframpError
-	InvalidMessage_CcipOfframpError
-	ReachedMaxSequenceNumber_CcipOfframpError
-	ManualExecutionNotAllowed_CcipOfframpError
-	InvalidInputsNumberOfAccounts_CcipOfframpError
-	InvalidInputsGlobalStateAccount_CcipOfframpError
-	InvalidInputsTokenIndices_CcipOfframpError
-	InvalidInputsPoolAccounts_CcipOfframpError
-	InvalidInputsTokenAccounts_CcipOfframpError
-	InvalidInputsSysvarAccount_CcipOfframpError
-	InvalidInputsFeeQuoterAccount_CcipOfframpError
-	InvalidInputsAllowedOfframpAccount_CcipOfframpError
-	InvalidInputsConfigAccounts_CcipOfframpError
-	InvalidInputsTokenAdminRegistryAccounts_CcipOfframpError
-	InvalidInputsLookupTableAccounts_CcipOfframpError
-	InvalidInputsLookupTableAccountWritable_CcipOfframpError
-	OfframpReleaseMintBalanceMismatch_CcipOfframpError
-	OfframpInvalidDataLength_CcipOfframpError
-	StaleCommitReport_CcipOfframpError
-	InvalidWritabilityBitmap_CcipOfframpError
-	InvalidCodeVersion_CcipOfframpError
-)
-
-func (value CcipOfframpError) String() string {
-	switch value {
-	case InvalidSequenceInterval_CcipOfframpError:
-		return "InvalidSequenceInterval"
-	case RootNotCommitted_CcipOfframpError:
-		return "RootNotCommitted"
-	case ExistingMerkleRoot_CcipOfframpError:
-		return "ExistingMerkleRoot"
-	case Unauthorized_CcipOfframpError:
-		return "Unauthorized"
-	case InvalidNonce_CcipOfframpError:
-		return "InvalidNonce"
-	case InvalidInputsMissingWritable_CcipOfframpError:
-		return "InvalidInputsMissingWritable"
-	case OnrampNotConfigured_CcipOfframpError:
-		return "OnrampNotConfigured"
-	case FailedToDeserializeReport_CcipOfframpError:
-		return "FailedToDeserializeReport"
-	case InvalidPluginType_CcipOfframpError:
-		return "InvalidPluginType"
-	case InvalidVersion_CcipOfframpError:
-		return "InvalidVersion"
-	case MissingExpectedPriceUpdates_CcipOfframpError:
-		return "MissingExpectedPriceUpdates"
-	case MissingExpectedMerkleRoot_CcipOfframpError:
-		return "MissingExpectedMerkleRoot"
-	case UnexpectedMerkleRoot_CcipOfframpError:
-		return "UnexpectedMerkleRoot"
-	case RedundantOwnerProposal_CcipOfframpError:
-		return "RedundantOwnerProposal"
-	case UnsupportedSourceChainSelector_CcipOfframpError:
-		return "UnsupportedSourceChainSelector"
-	case UnsupportedDestinationChainSelector_CcipOfframpError:
-		return "UnsupportedDestinationChainSelector"
-	case InvalidProof_CcipOfframpError:
-		return "InvalidProof"
-	case InvalidMessage_CcipOfframpError:
-		return "InvalidMessage"
-	case ReachedMaxSequenceNumber_CcipOfframpError:
-		return "ReachedMaxSequenceNumber"
-	case ManualExecutionNotAllowed_CcipOfframpError:
-		return "ManualExecutionNotAllowed"
-	case InvalidInputsNumberOfAccounts_CcipOfframpError:
-		return "InvalidInputsNumberOfAccounts"
-	case InvalidInputsGlobalStateAccount_CcipOfframpError:
-		return "InvalidInputsGlobalStateAccount"
-	case InvalidInputsTokenIndices_CcipOfframpError:
-		return "InvalidInputsTokenIndices"
-	case InvalidInputsPoolAccounts_CcipOfframpError:
-		return "InvalidInputsPoolAccounts"
-	case InvalidInputsTokenAccounts_CcipOfframpError:
-		return "InvalidInputsTokenAccounts"
-	case InvalidInputsSysvarAccount_CcipOfframpError:
-		return "InvalidInputsSysvarAccount"
-	case InvalidInputsFeeQuoterAccount_CcipOfframpError:
-		return "InvalidInputsFeeQuoterAccount"
-	case InvalidInputsAllowedOfframpAccount_CcipOfframpError:
-		return "InvalidInputsAllowedOfframpAccount"
-	case InvalidInputsConfigAccounts_CcipOfframpError:
-		return "InvalidInputsConfigAccounts"
-	case InvalidInputsTokenAdminRegistryAccounts_CcipOfframpError:
-		return "InvalidInputsTokenAdminRegistryAccounts"
-	case InvalidInputsLookupTableAccounts_CcipOfframpError:
-		return "InvalidInputsLookupTableAccounts"
-	case InvalidInputsLookupTableAccountWritable_CcipOfframpError:
-		return "InvalidInputsLookupTableAccountWritable"
-	case OfframpReleaseMintBalanceMismatch_CcipOfframpError:
-		return "OfframpReleaseMintBalanceMismatch"
-	case OfframpInvalidDataLength_CcipOfframpError:
-		return "OfframpInvalidDataLength"
-	case StaleCommitReport_CcipOfframpError:
-		return "StaleCommitReport"
-	case InvalidWritabilityBitmap_CcipOfframpError:
-		return "InvalidWritabilityBitmap"
-	case InvalidCodeVersion_CcipOfframpError:
-		return "InvalidCodeVersion"
-	default:
-		return ""
-	}
-}
