+++ chains/solana/gobindings/ccip_router/types.go
@@ -496,93 +496,3 @@ func (value RestoreOnAction) String() string {
 		return ""
 	}
 }
-
-type CcipRouterError ag_binary.BorshEnum
-
-const (
-	Unauthorized_CcipRouterError CcipRouterError = iota
-	InvalidInputsMint_CcipRouterError
-	InvalidVersion_CcipRouterError
-	FeeTokenMismatch_CcipRouterError
-	RedundantOwnerProposal_CcipRouterError
-	ReachedMaxSequenceNumber_CcipRouterError
-	InvalidInputsTokenIndices_CcipRouterError
-	InvalidInputsPoolAccounts_CcipRouterError
-	InvalidInputsTokenAccounts_CcipRouterError
-	InvalidInputsConfigAccounts_CcipRouterError
-	InvalidInputsTokenAdminRegistryAccounts_CcipRouterError
-	InvalidInputsLookupTableAccounts_CcipRouterError
-	InvalidInputsLookupTableAccountWritable_CcipRouterError
-	InvalidInputsTokenAmount_CcipRouterError
-	InvalidInputsTransferAllAmount_CcipRouterError
-	InvalidInputsAtaAddress_CcipRouterError
-	InvalidInputsAtaWritable_CcipRouterError
-	InvalidInputsChainSelector_CcipRouterError
-	InsufficientLamports_CcipRouterError
-	InsufficientFunds_CcipRouterError
-	SourceTokenDataTooLarge_CcipRouterError
-	InvalidTokenAdminRegistryInputsZeroAddress_CcipRouterError
-	InvalidTokenAdminRegistryProposedAdmin_CcipRouterError
-	SenderNotAllowed_CcipRouterError
-	InvalidCodeVersion_CcipRouterError
-	InvalidCcipVersionRollback_CcipRouterError
-)
-
-func (value CcipRouterError) String() string {
-	switch value {
-	case Unauthorized_CcipRouterError:
-		return "Unauthorized"
-	case InvalidInputsMint_CcipRouterError:
-		return "InvalidInputsMint"
-	case InvalidVersion_CcipRouterError:
-		return "InvalidVersion"
-	case FeeTokenMismatch_CcipRouterError:
-		return "FeeTokenMismatch"
-	case RedundantOwnerProposal_CcipRouterError:
-		return "RedundantOwnerProposal"
-	case ReachedMaxSequenceNumber_CcipRouterError:
-		return "ReachedMaxSequenceNumber"
-	case InvalidInputsTokenIndices_CcipRouterError:
-		return "InvalidInputsTokenIndices"
-	case InvalidInputsPoolAccounts_CcipRouterError:
-		return "InvalidInputsPoolAccounts"
-	case InvalidInputsTokenAccounts_CcipRouterError:
-		return "InvalidInputsTokenAccounts"
-	case InvalidInputsConfigAccounts_CcipRouterError:
-		return "InvalidInputsConfigAccounts"
-	case InvalidInputsTokenAdminRegistryAccounts_CcipRouterError:
-		return "InvalidInputsTokenAdminRegistryAccounts"
-	case InvalidInputsLookupTableAccounts_CcipRouterError:
-		return "InvalidInputsLookupTableAccounts"
-	case InvalidInputsLookupTableAccountWritable_CcipRouterError:
-		return "InvalidInputsLookupTableAccountWritable"
-	case InvalidInputsTokenAmount_CcipRouterError:
-		return "InvalidInputsTokenAmount"
-	case InvalidInputsTransferAllAmount_CcipRouterError:
-		return "InvalidInputsTransferAllAmount"
-	case InvalidInputsAtaAddress_CcipRouterError:
-		return "InvalidInputsAtaAddress"
-	case InvalidInputsAtaWritable_CcipRouterError:
-		return "InvalidInputsAtaWritable"
-	case InvalidInputsChainSelector_CcipRouterError:
-		return "InvalidInputsChainSelector"
-	case InsufficientLamports_CcipRouterError:
-		return "InsufficientLamports"
-	case InsufficientFunds_CcipRouterError:
-		return "InsufficientFunds"
-	case SourceTokenDataTooLarge_CcipRouterError:
-		return "SourceTokenDataTooLarge"
-	case InvalidTokenAdminRegistryInputsZeroAddress_CcipRouterError:
-		return "InvalidTokenAdminRegistryInputsZeroAddress"
-	case InvalidTokenAdminRegistryProposedAdmin_CcipRouterError:
-		return "InvalidTokenAdminRegistryProposedAdmin"
-	case SenderNotAllowed_CcipRouterError:
-		return "SenderNotAllowed"
-	case InvalidCodeVersion_CcipRouterError:
-		return "InvalidCodeVersion"
-	case InvalidCcipVersionRollback_CcipRouterError:
-		return "InvalidCcipVersionRollback"
-	default:
-		return ""
-	}
-}
