+++ chains/solana/gobindings/fee_quoter/types.go
@@ -849,126 +849,3 @@ func (value CodeVersion) String() string {
 		return ""
 	}
 }
-
-type FeeQuoterError ag_binary.BorshEnum
-
-const (
-	Unauthorized_FeeQuoterError FeeQuoterError = iota
-	InvalidInputs_FeeQuoterError
-	ZeroGasLimit_FeeQuoterError
-	DefaultGasLimitExceedsMaximum_FeeQuoterError
-	InvalidVersion_FeeQuoterError
-	RedundantOwnerProposal_FeeQuoterError
-	InvalidInputsMissingWritable_FeeQuoterError
-	InvalidInputsChainSelector_FeeQuoterError
-	InvalidInputsMint_FeeQuoterError
-	InvalidInputsMintOwner_FeeQuoterError
-	InvalidInputsTokenConfigAccount_FeeQuoterError
-	InvalidInputsMissingExtraArgs_FeeQuoterError
-	InvalidInputsMissingDataAfterExtraArgs_FeeQuoterError
-	InvalidInputsDestChainStateAccount_FeeQuoterError
-	InvalidInputsPerChainPerTokenConfig_FeeQuoterError
-	InvalidInputsBillingTokenConfig_FeeQuoterError
-	InvalidInputsAccountCount_FeeQuoterError
-	InvalidInputsNoUpdates_FeeQuoterError
-	InvalidInputsTokenAccounts_FeeQuoterError
-	DestinationChainDisabled_FeeQuoterError
-	FeeTokenDisabled_FeeQuoterError
-	MessageTooLarge_FeeQuoterError
-	UnsupportedNumberOfTokens_FeeQuoterError
-	InvalidEVMAddress_FeeQuoterError
-	InvalidEncoding_FeeQuoterError
-	InvalidTokenPrice_FeeQuoterError
-	StaleGasPrice_FeeQuoterError
-	InvalidInputsMissingTokenConfig_FeeQuoterError
-	MessageFeeTooHigh_FeeQuoterError
-	MessageGasLimitTooHigh_FeeQuoterError
-	ExtraArgOutOfOrderExecutionMustBeTrue_FeeQuoterError
-	InvalidExtraArgsTag_FeeQuoterError
-	InvalidChainFamilySelector_FeeQuoterError
-	InvalidTokenReceiver_FeeQuoterError
-	InvalidSVMAddress_FeeQuoterError
-	UnauthorizedPriceUpdater_FeeQuoterError
-	InvalidCodeVersion_FeeQuoterError
-)
-
-func (value FeeQuoterError) String() string {
-	switch value {
-	case Unauthorized_FeeQuoterError:
-		return "Unauthorized"
-	case InvalidInputs_FeeQuoterError:
-		return "InvalidInputs"
-	case ZeroGasLimit_FeeQuoterError:
-		return "ZeroGasLimit"
-	case DefaultGasLimitExceedsMaximum_FeeQuoterError:
-		return "DefaultGasLimitExceedsMaximum"
-	case InvalidVersion_FeeQuoterError:
-		return "InvalidVersion"
-	case RedundantOwnerProposal_FeeQuoterError:
-		return "RedundantOwnerProposal"
-	case InvalidInputsMissingWritable_FeeQuoterError:
-		return "InvalidInputsMissingWritable"
-	case InvalidInputsChainSelector_FeeQuoterError:
-		return "InvalidInputsChainSelector"
-	case InvalidInputsMint_FeeQuoterError:
-		return "InvalidInputsMint"
-	case InvalidInputsMintOwner_FeeQuoterError:
-		return "InvalidInputsMintOwner"
-	case InvalidInputsTokenConfigAccount_FeeQuoterError:
-		return "InvalidInputsTokenConfigAccount"
-	case InvalidInputsMissingExtraArgs_FeeQuoterError:
-		return "InvalidInputsMissingExtraArgs"
-	case InvalidInputsMissingDataAfterExtraArgs_FeeQuoterError:
-		return "InvalidInputsMissingDataAfterExtraArgs"
-	case InvalidInputsDestChainStateAccount_FeeQuoterError:
-		return "InvalidInputsDestChainStateAccount"
-	case InvalidInputsPerChainPerTokenConfig_FeeQuoterError:
-		return "InvalidInputsPerChainPerTokenConfig"
-	case InvalidInputsBillingTokenConfig_FeeQuoterError:
-		return "InvalidInputsBillingTokenConfig"
-	case InvalidInputsAccountCount_FeeQuoterError:
-		return "InvalidInputsAccountCount"
-	case InvalidInputsNoUpdates_FeeQuoterError:
-		return "InvalidInputsNoUpdates"
-	case InvalidInputsTokenAccounts_FeeQuoterError:
-		return "InvalidInputsTokenAccounts"
-	case DestinationChainDisabled_FeeQuoterError:
-		return "DestinationChainDisabled"
-	case FeeTokenDisabled_FeeQuoterError:
-		return "FeeTokenDisabled"
-	case MessageTooLarge_FeeQuoterError:
-		return "MessageTooLarge"
-	case UnsupportedNumberOfTokens_FeeQuoterError:
-		return "UnsupportedNumberOfTokens"
-	case InvalidEVMAddress_FeeQuoterError:
-		return "InvalidEVMAddress"
-	case InvalidEncoding_FeeQuoterError:
-		return "InvalidEncoding"
-	case InvalidTokenPrice_FeeQuoterError:
-		return "InvalidTokenPrice"
-	case StaleGasPrice_FeeQuoterError:
-		return "StaleGasPrice"
-	case InvalidInputsMissingTokenConfig_FeeQuoterError:
-		return "InvalidInputsMissingTokenConfig"
-	case MessageFeeTooHigh_FeeQuoterError:
-		return "MessageFeeTooHigh"
-	case MessageGasLimitTooHigh_FeeQuoterError:
-		return "MessageGasLimitTooHigh"
-	case ExtraArgOutOfOrderExecutionMustBeTrue_FeeQuoterError:
-		return "ExtraArgOutOfOrderExecutionMustBeTrue"
-	case InvalidExtraArgsTag_FeeQuoterError:
-		return "InvalidExtraArgsTag"
-	case InvalidChainFamilySelector_FeeQuoterError:
-		return "InvalidChainFamilySelector"
-	case InvalidTokenReceiver_FeeQuoterError:
-		return "InvalidTokenReceiver"
-	case InvalidSVMAddress_FeeQuoterError:
-		return "InvalidSVMAddress"
-	case UnauthorizedPriceUpdater_FeeQuoterError:
-		return "UnauthorizedPriceUpdater"
-	case InvalidCodeVersion_FeeQuoterError:
-		return "InvalidCodeVersion"
-	default:
-		return ""
-	}
-}
